Move some code to utils module
