+++ gwhat/utils/math.py
@@ -1,0 +1,43 @@
+# -*- coding: utf-8 -*-
+
+# Copyright © 2014-2018 GWHAT Project Contributors
+# https://github.com/jnsebgosselin/gwhat
+#
+# This file is part of GWHAT (Ground-Water Hydrograph Analysis Toolbox).
+# Licensed under the terms of the GNU General Public License.
+
+import numpy as np
+import datetime
+
+
+def clip_time_series(tclip, tp, xp):
+    """
+    Clip tp and xp on tclip. tclip and tp must be arrays of numerical
+    Excel times. Return two empty arrays if tclip and tp are
+    mutually exclusive.
+    """
+    # Check that tp and tclip are not mutually exclusive.
+    if len(np.unique(np.hstack([tclip, tp]))) == (len(tclip) + len(tp)):
+        return [], []
+
+    # Clip xp and tp on tclip.
+    if tp[0] < tclip[0]:
+        idx = np.where(tp >= tclip[0])[0][0]
+        tp = tp[idx:]
+        xp = xp[idx:]
+    if tp[-1] > tclip[-1]:
+        idx = np.where(tp <= tclip[-1])[0][-1]
+        tp = tp[:idx+1]
+        xp = xp[:idx+1]
+    return tp, xp
+
+
+def convert_date_to_datetime(years, months, days):
+    """
+    Produce datetime series from years, months, and days series.
+    """
+    dates = [0] * len(years)
+    for t in range(len(years)):
+        dates[t] = datetime.datetime(
+                int(years[t]), int(months[t]), int(days[t]), 0)
+    return dates
